requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
baidu-aip==4.16.13
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.3.1
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
multidict==6.0.5
openai==0.28.0
PyAudio==0.2.14
pydantic==2.6.3
pydantic_core==2.16.3
pygame==2.5.2
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.31.0
sniffio==1.3.1
SpeechRecognition==3.10.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.7.0
yarl==1.9.4
